requirements.txt
azure-functions
azure-storage-blob
unstructured[all-docs]
python-magic